pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexarq-cli"
version = "0.1.0"
description = "multi-agent code review CLI powered by LLMs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["code-review", "ai", "llm", "security", "git", "multi-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Security",
]

dependencies = [
    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",
    # Git
    "gitpython>=3.1.40",
    # Async HTTP
    "httpx>=0.27.0",
    # LLM providers
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "google-genai>=1.0.0",
    "ollama>=0.3.0",
    # Config / validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    # Security
    "keyring>=25.2.0",
    "cryptography>=42.0.0",
    # Reliability
    "tenacity>=8.3.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "crewai>=0.63.0",
    "pyautogen>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
]

[project.scripts]
nexarq         = "nexarq_cli.main:app"
nexarq-setup   = "nexarq_cli.cli.install_cmd:_post_install"

[project.urls]
Homepage = "https://github.com/nexarq-cli"
Repository = "https://github.com/nexarq-cli"
Issues = "https://github.com/nexarq-cli/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["nexarq_cli*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "S", "B"]
ignore = ["S101"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
